Rename NodePanel -> NodeProvider
